try catch agregado

+++ admin/habitaciones/eliminar_habitacion.php
@@ -1,12 +1,19 @@
 <?php
 require('../../db.php');
 
+if(isset($_GET['id_habitaciones'])){
+
 $id_habitacion = $_GET['id_habitaciones'];
 
 $conexiondb = conectardb();
 
-$query ="DELETE FROM habitaciones WHERE id_habitaciones=".$id_habitacion;
-
+try {
+    $query ="DELETE FROM habitaciones WHERE id_habitaciones=".$id_habitacion;
+}
+finally{
+    echo "<script>alert('Registros Actualizados');
+window.location.href='../listado/form_habitaciones.php'</script>";
+}
 $respuesta= mysqli_query($conexiondb, $query);
 
 if ($respuesta) {
@@ -15,6 +22,10 @@ if ($respuesta) {
 } else {
     echo "<script>alert('No se elimino el registro');
     window.location.href='../listado/form_habitaciones.php'</script>";
+}
+}else{
+echo "<script>alert('No se elimino el registro');
+window.location.href='../listado/form_habitaciones.php'</script>";
 }
 
 ?>
